requirements.txt
adal==1.2.7
applicationinsights==0.11.10
astroid==2.8.4
azure-common==1.1.27
azure-core==1.19.0
azure-graphrbac==0.61.1
azure-identity==1.4.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==8.1.0
azure-mgmt-core==1.3.0
azure-mgmt-keyvault==9.1.0
azure-mgmt-resource==13.0.0
azure-mgmt-storage==11.2.0
azure-storage-queue==12.1.6
azureml-automl-core==1.35.0
azureml-automl-runtime==1.34.1.post1
azureml-core==1.35.0
azureml-dataprep==2.23.2
azureml-dataprep-native==38.0.0
azureml-dataprep-rslex==1.21.2
azureml-dataset-runtime==1.35.0
azureml-defaults==1.34.0
azureml-inference-server-http==0.3.2
azureml-interpret==1.34.0
azureml-pipeline==1.35.0
azureml-pipeline-core==1.35.0
azureml-pipeline-steps==1.35.0
azureml-telemetry==1.35.0
azureml-train-automl==1.34.0
azureml-train-automl-client==1.35.0
azureml-train-automl-runtime==1.34.1
azureml-train-core==1.35.0
azureml-train-restclients-hyperdrive==1.35.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bokeh==2.3.3
boto==2.49.0
boto3==1.15.18
botocore==1.18.18
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
cloudpickle==1.6.0
configparser==3.7.4
contextlib2==21.6.0
contextvars==2.4
cryptography==3.4.8
cycler==0.10.0
Cython==0.29.24
dask==2.30.0
dill==0.3.4
distributed==2.30.1
distro==1.6.0
docker==5.0.2
dotnetcore2==2.1.21
fire==0.4.0
Flask==1.0.3
flatbuffers==2.0
fsspec==2021.10.0
fusepy==3.0.1
gensim==3.8.3
google-api-core==2.0.1
google-auth==2.2.1
googleapis-common-protos==1.53.0
gunicorn==20.1.0
HeapDict==1.0.1
idna==3.2
immutables==0.16
importlib-metadata==4.8.1
inference-schema==1.3.0
interpret-community==0.19.3
interpret-core==0.2.5
isodate==0.6.0
isort==5.9.3
itsdangerous==2.0.1
jeepney==0.7.1
Jinja2==3.0.2
jmespath==0.10.0
joblib==0.14.1
json-logging-py==0.2
jsonpickle==2.0.0
keras2onnx==1.6.0
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
lightgbm==2.3.0
llvmlite==0.36.0
locket==0.2.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mccabe==0.6.1
msal==1.15.0
msal-extensions==0.2.2
msgpack==1.0.2
msrest==0.6.21
msrestazure==0.6.4
ndg-httpsclient==0.5.1
nimbusml==1.8.0
numba==0.53.1
numpy==1.18.5
oauthlib==3.1.1
onnx==1.7.0
onnxconverter-common==1.6.0
onnxmltools==1.4.1
onnxruntime==1.8.0
opencensus==0.8.0
opencensus-context==0.1.2
opencensus-ext-azure==1.0.8
packaging==21.0
pandas==0.25.3
partd==1.2.0
pathspec==0.9.0
patsy==0.5.2
Pillow==8.3.2
platformdirs==2.4.0
pmdarima==1.1.1
portalocker==1.7.1
protobuf==3.18.0
psutil==5.8.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pylint==2.11.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.3.7
scikit-learn==0.22.2.post1
scipy==1.5.2
SecretStorage==3.3.1
shap==0.39.0
six==1.16.0
skl2onnx==1.4.9
sklearn-pandas==1.7.0
slicer==0.0.7
smart-open==1.9.0
sortedcontainers==2.4.0
statsmodels==0.10.2
tblib==1.7.0
termcolor==1.1.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.62.3
typed-ast==1.4.3
typing-extensions==3.10.0.2
urllib3==1.26.7
websocket-client==1.2.1
Werkzeug==1.0.1
wrapt==1.12.1
xgboost==0.90
zict==2.0.0
zipp==3.6.0
